+++ backend/src/controllers/StudentController.ts
@@ -8,8 +8,8 @@ import GetLanguagesService from '../services/GetLanguagesService';
 import CreateStudentService from '../services/CreateStudentService';
 import CreateRepositoryService from '../services/CreateRepositoryService';
 import GetPeriodStudentDailyReportsService from '../services/GetPeriodStudentDailyReportsService';
-import GetClassReport from '../services/GetClassReport';
 import Student from '../models/Student';
+import AppError from '../errors/AppError';
 
 class StudentController {
   static async store(request: Request, response: Response): Promise<Response> {
@@ -67,80 +67,30 @@ class StudentController {
     until = until || new Date().toISOString();
     const { since } = request.query;
 
-    if (username) {
-      const getUser = new GetUserService();
-      const { github_id } = await getUser.execute({
-        username,
-      });
-
-      const studentsRepository = getRepository(Student);
-      const student = await studentsRepository.findOne({
-        where: {
-          github_id,
-        },
-      });
-
-      const getPeriodStudentDailyReport = new GetPeriodStudentDailyReportsService();
-      const studentDailyReport = await getPeriodStudentDailyReport.execute(
-        student.id,
-        String(since),
-        String(until),
-      );
-
-      return response.json(studentDailyReport);
+    if (!username) {
+      throw new AppError('Github username in params is required.', 400);
     }
 
-    const getClassReport = new GetClassReport();
-    const classInformation = await getClassReport.execute(
+    const getUser = new GetUserService();
+    const { github_id } = await getUser.execute({
+      username,
+    });
+
+    const studentsRepository = getRepository(Student);
+    const student = await studentsRepository.findOne({
+      where: {
+        github_id,
+      },
+    });
+
+    const getPeriodStudentDailyReport = new GetPeriodStudentDailyReportsService();
+    const studentDailyReport = await getPeriodStudentDailyReport.execute(
+      student.id,
       String(since),
       String(until),
     );
-    return response.json(classInformation);
-  }
 
-  static async index(request: Request, response: Response): Promise<Response> {
-    const studentsRepository = getRepository(Student);
-    const students = await studentsRepository.find();
-
-    if (request.path === '/below_average') {
-      const schema = yup.object().shape({
-        since: yup.string().required('Since date string is required.'),
-        until: yup.string(),
-      });
-      await schema.validate(request.query);
-
-      let { until } = request.query;
-      until = until || new Date().toISOString();
-      const { since } = request.query;
-
-      const getClassReport = new GetClassReport();
-      const classInformation = await getClassReport.execute(
-        String(since),
-        String(until),
-      );
-
-      const belowAverageStudents: Student[] = [];
-
-      for (const student of students) {
-        const getPeriodStudentDailyReport = new GetPeriodStudentDailyReportsService();
-        const studentDailyReport = await getPeriodStudentDailyReport.execute(
-          student.id,
-          String(since),
-          String(until),
-        );
-
-        if (
-          studentDailyReport.new_interactions <
-          classInformation.new_interactions_average
-        ) {
-          belowAverageStudents.push(student);
-        }
-      }
-
-      return response.json(belowAverageStudents);
-    }
-
-    return response.json(students);
+    return response.json(studentDailyReport);
   }
 }
