+++ backend/src/routes/index.ts
@@ -3,7 +3,8 @@ import { Router } from 'express';
 import authMiddleware from '../middlewares/authMiddleware';
 import sessionRouter from './session.routes';
 import teacherRouter from './teacher.routes';
-import studentsRouter from './students.routes';
+import studentRouter from './student.routes';
+import classRouter from './class.routes';
 import userRouter from './user.routes';
 
 const routes = Router();
@@ -14,6 +15,7 @@ routes.use('/session', sessionRouter);
 
 routes.use(authMiddleware);
 
-routes.use('/students', studentsRouter);
+routes.use('/student', studentRouter);
+routes.use('/class', classRouter);
 
 export default routes;
