+++ backend/src/routes/student.routes.ts
@@ -1,0 +1,10 @@
+import { Router } from 'express';
+
+import StudentController from '../controllers/StudentController';
+
+const studentRouter = Router();
+
+studentRouter.post('/', StudentController.store);
+studentRouter.get('/report/:username', StudentController.show);
+
+export default studentRouter;
